subscriber/requirements.txt
# This file defines the image's dependencies
redis
google-cloud-pubsub
Flask



